// src/app/interfaces/gesaca.ts
/* tslint:disable */


/**
 * AUTO-GENERATED FILE @ 2019-07-25 13:41:46 - DO NOT EDIT!
 *
 * This file was automatically generated by schemats v.3.0.3
 * $ schemats generate -c mysql://username:password@localhost/db_gesaca -C -t anio -t matricula -t nivel -t persona -s db_gesaca
 *
 */


export namespace AnioFields {
    export type idAnio = number;
    export type fecha = number;
    export type fechaInicio = Date;
    export type fechaFin = Date;
    export type descripcion = string | null;

}

export interface Anio {
    IdAnio: AnioFields.idAnio;
    Fecha: AnioFields.fecha;
    FechaInicio: AnioFields.fechaInicio;
    FechaFin: AnioFields.fechaFin;
    Descripcion: AnioFields.descripcion;

}

export namespace MatriculaFields {
    export type idMatricula = number;
    export type idPersona = number;
    export type idNivel = number;
    export type idAnio = number;
    export type grado = number;
    export type seccion = number;
    export type nota = number | null;

}

export interface Matricula {
    IdMatricula: MatriculaFields.idMatricula;
    IdPersona: MatriculaFields.idPersona;
    IdNivel: MatriculaFields.idNivel;
    IdAnio: MatriculaFields.idAnio;
    Grado: MatriculaFields.grado;
    Seccion: MatriculaFields.seccion;
    Nota: MatriculaFields.nota;

}

export namespace NivelFields {
    export type idNivel = number;
    export type descripcion = string;
    export type estado = number;

}

export interface Nivel {
    IdNivel: NivelFields.idNivel;
    Descripcion: NivelFields.descripcion;
    Estado: NivelFields.estado;

}

export namespace PersonaFields {
    export type idPersona = number;
    export type dni = string;
    export type nombre = string;
    export type paterno = string;
    export type materno = string;
    export type telefono = string | null;
    export type tipo = number;
    export type sub = number;

}

export interface Persona {
    IdPersona: PersonaFields.idPersona;
    Dni: PersonaFields.dni;
    Nombre: PersonaFields.nombre;
    Paterno: PersonaFields.paterno;
    Materno: PersonaFields.materno;
    Telefono: PersonaFields.telefono;
    Tipo: PersonaFields.tipo;
    Sub: PersonaFields.sub;

}